pyproject.toml
[tool.poetry]
name = "mit-skate-cal"
version = "0.1.0"
description = ""
authors = ["Max Katz-Christy <maxkatzchristy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
caldav = "^0.8.0"
camelot-py = {extras = ["base"], version = "^0.10.1"}

[tool.poetry.dev-dependencies]
matplotlib = "^3.5.0"
pyvoronoi = "^1.0.7"
Shapely = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
